backend/requirements.txt
Flask==3.0.3
gunicorn==22.0.0

Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

Flask-JWT-Extended==4.6.0
passlib==1.7.4

marshmallow==3.21.3
python-dotenv==1.0.1

flask-smorest==0.44.0

qrcode[pil]==7.4.2
Pillow==10.4.0
